# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hub William API",
    "description": "Control-plane and gateway API for Hub William agent pools",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/agent-connections/{connection_id}": {
      "get": {
        "tags": ["agent connections"],
        "operationId": "get_connection",
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "description": "Agent connection ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Latest provider connection status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentConnection"
                }
              }
            }
          },
          "404": {
            "description": "Connection not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["agent connections"],
        "operationId": "disconnect",
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "description": "Agent connection ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Connected account pool deleted"
          },
          "404": {
            "description": "Connection not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentAuthorizationPrompt": {
        "type": "object",
        "required": [
          "authorization_url",
          "expires_at",
          "poll_after_seconds",
          "requires_callback_url"
        ],
        "properties": {
          "authorization_url": {
            "type": "string"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "poll_after_seconds": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "requires_callback_url": {
            "type": "boolean"
          },
          "user_code": {
            "type": ["string", "null"]
          }
        }
      },
      "AgentConnection": {
        "type": "object",
        "required": [
          "availability_status",
          "created_at",
          "id",
          "provider",
          "status",
          "updated_at"
        ],
        "properties": {
          "account_label": {
            "type": ["string", "null"]
          },
          "authorization": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AgentAuthorizationPrompt"
              }
            ]
          },
          "availability_status": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "failure_message": {
            "type": ["string", "null"]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "plan": {
            "type": ["string", "null"]
          },
          "provider": {
            "$ref": "#/components/schemas/AgentProvider"
          },
          "status": {
            "$ref": "#/components/schemas/AgentConnectionStatus"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AgentConnectionStatus": {
        "type": "string",
        "enum": ["pending", "connected", "failed", "disconnected"]
      },
...
      "AgentProvider": {
        "type": "string",
        "enum": ["chatgpt", "claude", "gemini", "grok", "deepseek", "groq"]
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": ["code", "message"],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}